Moved IScreenInfoUiaProvider into shared space. Both projects must now implement the virtual functions. ConHost should be MOSTLY set up. WT needs a way to implement these (need access to Terminal info)
NEXT: Fix IScreenInfoUiaProvider's references to Selection and UiaTextRange
